Fix bug when handling deleted documents. Fix bug in HTML harvester. Add 'api' module.
From the it-is-friday-i-ll-make-a-single-commit-for-everything dep't.

+++ tests/test_whoosh.py
@@ -5,13 +5,13 @@ from __future__ import unicode_literals
 
 import os
 import shutil
-import time
 try:  # Python 2.6 compatibility
     from unittest2 import TestCase
 except ImportError:
     from unittest import TestCase
 
-from dokang.backends import whoosh as whoosh_backend
+from dokang import api
+from dokang.harvesters import html_config
 
 
 def get_data_path(*components):
@@ -21,43 +21,43 @@ def get_data_path(*components):
         *components)
 
 
-class TestWhooshBackend(TestCase):
+class TestApi(TestCase):
 
     def tearDown(self):
-        super(TestWhooshBackend, self).tearDown()
-        shutil.rmtree(self.index_path)
+        super(TestCase, self).tearDown()
+        if os.path.exists(self.index_path):
+            shutil.rmtree(self.index_path)
 
     @property
     def index_path(self):
         return get_data_path('whoosh_test_index')
 
     def test_integration(self):
-        indexer = whoosh_backend.WhooshIndexer(self.index_path)
-        indexer.initialize()
-        docs = [{
-            'title': 'TitleOfFoo',
-            'content': 'ContentOfFoo',
-            'path': 'foo.html',
-            'mtime': time.time(),
-            'set': 'set1',
-            'kind': 'HTML'}]
-        indexer.index_documents(docs)
+        api.initialize_index(self.index_path)
 
-        searcher = whoosh_backend.WhooshSearcher(self.index_path)
-        results = list(searcher.search('NotIndexed'))
-        self.assertEqual(len(results), 0)
+        doc_set_info = {
+            'id': 'test',
+            'title': 'Test documentation',
+            'path': get_data_path('api'),
+            'url': 'http://docs.exemple.com/',
+            'harvester': html_config()
+        }
+        api.index_document_set(self.index_path, doc_set_info)
 
-        results = list(searcher.search('ContentOfFoo'))
-        self.assertEqual(len(results), 1)
-        self.assertEqual(results[0]['path'], 'foo.html')
-        self.assertEqual(results[0]['title'], 'TitleOfFoo')
-        self.assertEqual(results[0]['set'], 'set1')
+        self.assertEqual(list(api.search(self.index_path, "ShouldNotBeIndexed")), [])
+        hits = list(api.search(self.index_path, "ShouldBeIndexed"))
+        self.assertEqual(len(hits), 1)
+        self.assertEqual(hits[0]['path'], 'test1.html')
+        self.assertEqual(hits[0]['title'], 'The title')
+        self.assertEqual(hits[0]['set'], 'test')
+        self.assertEqual(hits[0]['kind'], 'HTML')
+        hits = list(api.search(self.index_path, "ShouldBeIndexed set:test"))
+        self.assertEqual(len(hits), 1)
+        self.assertEqual(list(api.search(self.index_path, "ShouldBeIndexed set:unknown")), [])
 
-        # Change the content of foo, reindex and make sure that we
-        # cannot find the old content.
-        docs[0]['content'] = 'NewContentOfFoo'
-        indexer.index_documents(docs)
-        results = list(searcher.search('ContentOfFoo'))
-        self.assertEqual(len(results), 0)
-        results = list(searcher.search('NewContentOfFoo'))
-        self.assertEqual(len(results), 1)
+        api.clear_document_set(self.index_path, 'unknown')
+        hits = list(api.search(self.index_path, "ShouldBeIndexed"))
+        self.assertEqual(len(hits), 1)
+
+        api.clear_document_set(self.index_path, doc_set_info['id'])
+        self.assertEqual(list(api.search(self.index_path, "ShouldBeIndexed")), [])
